+++ icmake/icm-comp/symtab/funaddress.c
@@ -1,6 +1,6 @@
 #include "symtab.ih"
 
-uint16_t symtab_funAddress(size_t funIdx)
+uint16_t symtab_funAddress(unsigned funIdx)
 {
     return gs_functions.info[funIdx].address;
 }
