+++ icmake/icm-comp/symtab/funparametertype.c
@@ -1,6 +1,6 @@
 #include "symtab.ih"
 
-ExprType symtab_funParameterType(size_t funIdx, size_t paramIdx)
+ExprType symtab_funParameterType(unsigned funIdx, unsigned paramIdx)
 {
                                             /* &typeMask removes e_reg */
     return gs_functions.info[funIdx].paramType[paramIdx] & e_typeMask;
