+++ icmake/icm-comp/symtab/nlocals.c
@@ -1,6 +1,6 @@
 #include "symtab.ih"
 
-size_t symtab_nLocals()
+unsigned symtab_nLocals()
 {
     return gs_vars.varTab[1].nVars - gs_vars.nParams;
 }
