+++ icmake/icm-comp/symtab/nparams.c
@@ -1,6 +1,6 @@
 #include "symtab.ih"
 
-size_t symtab_nParams()
+unsigned symtab_nParams()
 {
     return gs_vars.nParams;
 }
