+++ icmake/icm-comp/symtab/stcomparevars.c
@@ -3,6 +3,6 @@
 int st_compareVars(void const *key, void const *idxPtr)
 {
     return strcmp((char const *)key, 
-                    gs_compareTable->info[*(size_t const *)idxPtr].name);
+                    gs_compareTable->info[*(unsigned const *)idxPtr].name);
 }
