+++ icmake/icm-comp/symtab/stnestingoffset.c
@@ -1,6 +1,6 @@
 #include "symtab.ih"
 
-size_t st_nestingOffset()
+unsigned st_nestingOffset()
 {
     return gs_vars.nLevels - 1;
 }
