+++ icmake/icm-comp/symtab/symtab.h
@@ -27,26 +27,26 @@ int         symtab_addFunction(ExprType type);
 int         symtab_findFun(void); /* util_string() holds the name of a function 
                                 returns its idx or -1 if not found       */
 
-size_t      symtab_fun_nParams(size_t funIdx);
+unsigned      symtab_fun_nParams(unsigned funIdx);
 
 void        symtab_newParam();  /* counts the number of params so far */
 
-char const *symtab_funName(size_t funIdx); 
+char const *symtab_funName(unsigned funIdx); 
 
-uint16_t       symtab_funAddress(size_t funIdx); 
+uint16_t       symtab_funAddress(unsigned funIdx); 
 
-ExprType    symtab_funType(size_t funIdx);
+ExprType    symtab_funType(unsigned funIdx);
 
-ExprType    symtab_funParameterType(size_t funIdx, size_t paramIdx);
+ExprType    symtab_funParameterType(unsigned funIdx, unsigned paramIdx);
 
-void        symtab_writeFunAddress(FILE *bin, size_t funidx);
+void        symtab_writeFunAddress(FILE *bin, unsigned funidx);
 
-size_t      symtab_lastFunction(void);  /* index of last defined function */
+unsigned      symtab_lastFunction(void);  /* index of last defined function */
 
 void        symtab_setFunParams();      /* set nparams & types of the function
                                             that's currently being defined */
 
-size_t      symtab_setFunctionInfo(size_t nParams); /* initializes the fields 
+unsigned      symtab_setFunctionInfo(unsigned nParams); /* initializes the fields 
                                                  of the last defined function;
                                                  returns # local variables */
 
@@ -58,18 +58,18 @@ VarIndex    symtab_defineVar(ExprType type); /* Define a var. at the topmost
                                                 already def'd) */
 ExprType    symtab_varType(VarIndex vi); /* type of variable vi */
 
-size_t      symtab_nGlobals(void);   /* the number of global variables defined
+unsigned      symtab_nGlobals(void);   /* the number of global variables defined
                                         so far */
-size_t      symtab_nLocals(void);    /* the number of local variables of the 
+unsigned      symtab_nLocals(void);    /* the number of local variables of the 
                                         function currently being defined */
 
-ExprType    symtab_localType(size_t idx);   /* type of local variable idx, 1st 
+ExprType    symtab_localType(unsigned idx);   /* type of local variable idx, 1st 
                                                 local var has idx 0 */
 
 void        symtab_writeGlobals(FILE *bin);
 
 void        symtab_setNparams(void);
-size_t      symtab_nParams();
+unsigned      symtab_nParams();
 
 void        symtab_cleanup(void);       /* pop all local variables */
