+++ icmake/icm-comp/symtab/writefunaddress.c
@@ -1,6 +1,6 @@
 #include "symtab.ih"
 
-void symtab_writeFunAddress(FILE *bin, size_t funIdx)
+void symtab_writeFunAddress(FILE *bin, unsigned funIdx)
 {
     util_out(bin, &gs_functions.info[funIdx].address, sizeof(uint16_t));
 }
