+++ icmake/icm-comp/util/util.h
@@ -9,7 +9,7 @@ char const *util_string(void);
 int         util_printf(char const *fmt, ...);  /* fill gu_lexstring */
 
 void        util_catString(char const *txt);
-void        util_out(FILE *bin, void const *source, size_t sourceSizeInBytes);
+void        util_out(FILE *bin, void const *source, unsigned sourceSizeInBytes);
 void        util_resetSemErr(void);
 void        util_semantic(char const *s, ...);
 void        util_setSourceName(char const *txt);
