+++ icmake/icm-exec/aux/set.c
@@ -1,6 +1,6 @@
 #include "aux.ih"
 
-void aux_set(size_t offset)
+void aux_set(unsigned offset)
 {
     ga_offset = offset;
 }
